IK-100 [Refactored] Moved handlepopulateInboxResponse from InboxAdapter.java to InboxActivity.java. Request is now handled by activity.
