+++ client/app/src/main/java/com/example/mrides/Activity/InboxActivity.java
@@ -12,21 +12,30 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.example.mrides.Notification.InboxAdapter;
+import com.example.mrides.Notification.Invitation;
 import com.example.mrides.Notification.MatchingMessagingService;
 import com.example.mrides.R;
 import com.example.mrides.controller.RequestHandler;
 import com.example.mrides.controller.Subject;
+import com.example.mrides.userDomain.PassengerSerializer;
 import com.google.firebase.messaging.RemoteMessage;
 
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
-public class InboxActivity extends AppCompatActivity {
+public class InboxActivity extends AppCompatActivity implements ActivityObserver{
 
     private RecyclerView mRecyclerView;
     private RecyclerView.Adapter mAdapter;
     private RecyclerView.LayoutManager mLayoutManager;
     private RemoteMessage notification;
     private RequestHandler requestHandler = new RequestHandler();
+    private List<Invitation> invitations = new ArrayList<>();
 
     /**
      * When the inbox is created the inbox gets a list of past notifications, inorder
@@ -47,11 +56,43 @@ public class InboxActivity extends AppCompatActivity {
         // use a linear layout manager
         mLayoutManager = new LinearLayoutManager(this);
         mRecyclerView.setLayoutManager(mLayoutManager);
-        mAdapter = new InboxAdapter(this);
+        //getInboxData(); TODO call this method to get the list of inbox data
+    }
+
+    private void getInboxData(){
+        requestHandler.attach(this);
+        Map<String,String> userInfo = PassengerSerializer.getParameters(RequestHandler.getUser());
+        requestHandler.httpPostStringRequest("http://"+this.getString(R.string.web_server_ip)+
+                        "/getUserInboxList.php",userInfo,
+                "application/x-www-form-urlencoded; charset=UTF-8", this);
+    }
+
+    @Override
+    public void Update(String response) {
+        handlepopulateInboxResponse(response);
+        mAdapter = new InboxAdapter(this,invitations);
         mRecyclerView.setAdapter(mAdapter);
         if(getIntent().getParcelableExtra("NOTIFICATION")!=null) {
             notification = (RemoteMessage) (getIntent().getParcelableExtra("NOTIFICATION"));
             ((InboxAdapter) mAdapter).setViewComponents(notification.getData());
         }
+    }
+
+    private void handlepopulateInboxResponse(String response){
+        try {
+            JSONObject inboxResponse = new JSONObject(response);
+            JSONArray invites = inboxResponse.getJSONArray("invites");
+            for(int index =0;index<invites.length();index++){
+                JSONObject inviteJson = invites.getJSONObject(index);
+                Invitation invite = new Invitation(inviteJson.getString("driverEmail"),
+                        inviteJson.getString("review"),inviteJson.getString("driverUrlPic"),
+                        inviteJson.getString("displayName"));
+                invitations.add(invite);
+            }
+
+        } catch (JSONException e) {
+            e.printStackTrace();
+        }
+
     }
 }
